MVP/pyproject.toml
[project]
name = "news-agent"
version = "0.1.0"
description = "新闻爬虫与 AI 整理系统"
requires-python = ">=3.9"
authors = [
    { name = "Team" }
]
dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "lxml>=4.9.0",
    "playwright>=1.30.0",
    "sqlalchemy>=1.4.0,<3.0.0",
    "apscheduler>=3.9.0",
    "openai>=1.0.0",
    "pydantic>=1.10.0,<3.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
    "yagmail>=0.15.0",
    "flask>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line-length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
